added driver code after wipeout

+++ sw/airborne/modules/optical_flow/px4flow.h
@@ -47,8 +47,8 @@ struct mavlink_optical_flow {
   float flow_comp_m_x;    ///< Flow in meters in x-sensor direction, angular-speed compensated [meters/sec]
   float flow_comp_m_y;    ///< Flow in meters in y-sensor direction, angular-speed compensated [meters/sec]
   float ground_distance;  ///< Ground distance in meters. Positive value: distance known. Negative value: Unknown distance
-  int16_t flow_x;         ///< Flow in pixels in x-sensor direction
-  int16_t flow_y;         ///< Flow in pixels in y-sensor direction
+  int32_t flow_x;         ///< Flow in pixels in x-sensor direction
+  int32_t flow_y;         ///< Flow in pixels in y-sensor direction
   uint8_t sensor_id;      ///< Sensor ID
   uint8_t quality;        ///< Optical flow quality / confidence. 0: bad, 255: maximum quality
 };
